feat: added npc units (#211)
* set up pseudo npc behavior tree
Todo: add ai unit that will use this tree
make bevaviors adjusted for this unit

* removed hard coded sphynx index value from behaviors.
Instead an index is to be set in the AI manager which the AI entity will use for its logic

* began targeting changes for AI so that it can target more than just the player

* Changed how the AI target live entities (player or other ai units)

* adding team logo

* Updated targeting behavior for enemy AI
added "todo:" tasks in enemy_behaviors.cpp

* Other ai unit being drawn
todo: needs scaling, behavior tree not working for some reason

* npc unit now roaming in the map

* NPC units can now be killed by the AI properly
added freeze_frame function used by the npc to freeze them when they die
added shadow blob to npc units

* added ai to second level

* added textures for broken door and npc
loaded the textures, added code to change textures to broken door when door is broken

* Changed roam check behavior so that it would not stop roaming where there was a dead body

+++ src/cpp/level_helper.cpp
@@ -430,7 +430,7 @@ namespace crow {
 		furn_matrix[3][1] = 1.f;
 
 		// this switch case handles the rotation of the matrix of the object in addition to its collision
-		switch (orientation) {
+		switch (orientation) { // TODO: CLEAN UP WARNINGS, USE A CLAMPING FUNCTION FOR INTS NOT FLOATS, THIS COULD BE DANGEROUS
 		case 'u': // up
 			furn_matrix = XrotationM(furn_matrix, 90);
 			for (int i = 0; i < 4; i++) {
@@ -475,7 +475,7 @@ namespace crow {
 		r.furniture_meshes.push_back(game_manager::mesh_types::BED1);
 		r.furniture_textures.push_back(game_manager::texture_list::BED1);
 
-		return r.furniture_matrices.size() - 1;
+		return (int)r.furniture_matrices.size() - 1;
 	}
 
 	int level::place_chair(room& r, float2e position, char orientation) {
@@ -508,8 +508,8 @@ namespace crow {
 			break;
 		}
 
-		// collision is the same for all orientations
-		r.tiles.map[clampf(position.y, 0, r.length - 1)][clampf(position.x, 0, r.width - 1)]->is_open = false;
+		// collision is the same for all orientations TODO: You may want to use a clamp function that takes in ints instead of float, this is not safe also throws warnings
+		r.tiles.map[clampf(position.y, 0.f, r.length - 1.f)][clampf(position.x, 0.f, r.width - 1.f)]->is_open = false;
 
 		// correctly set the furniture's position on the map
 		furn_matrix[3][0] += position.x - r.width * 0.5f + 0.5f; furn_matrix[3][2] += position.y - r.length * 0.5f + 0.5f;
